+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Package is prepared for npm publication as `@crcatala/ticktick-cli`, with `tt` and `ttcli` command aliases.
+
 ## [0.1.0] - 2024-12-17
 
 ### Added